issue: 1537043 Cleanup hw/mlx5 folder
This content is moved to qp_mgr_eth_mlx5.
