+++ src/vma/dev/qp_mgr_eth_mlx5.h
@@ -42,6 +42,16 @@
 
 #include <infiniband/mlx5_hw.h>
 
+
+struct mlx5_wqe64 {
+	union {
+		struct mlx5_wqe_ctrl_seg ctrl;
+		uint32_t data[4];
+	} ctrl;
+	struct mlx5_wqe_eth_seg eseg;
+	struct mlx5_wqe_data_seg dseg;
+};
+
 class qp_mgr_eth_mlx5 : public qp_mgr_eth
 {
 friend class cq_mgr_mlx5;
